+++ WPF/VMagicMirrorConfig/Depdency/ModelInstaller.cs
@@ -33,6 +33,11 @@
             resolver.Add(new SaveFileManager());
 
             resolver.Add(new AvatarLoader());
+
+            resolver.Add(new RuntimeHelper());
+            resolver.Add(new CustomMotionList());
+            resolver.Add(new ImageQualitySetting());
+            resolver.Add(new ExternalTrackerRuntimeConfig());
         }
     }
 }
