+++ WPF/VMagicMirrorConfig/Model/Avatar/AvatarLoader.cs
