+++ WPF/VMagicMirrorConfig/Model/Avatar/FaceBlendshapeNameStore.cs
@@ -1,0 +1,139 @@
+﻿using System;
+using System.Collections.ObjectModel;
+using System.Linq;
+
+namespace Baku.VMagicMirrorConfig
+{
+    //NOTE: ExTracker用ではないやつ。
+
+    /// <summary>
+    /// ブレンドシェイプクリップ名を一覧保持するクラスです。
+    /// ExTrackerとクラスが分かれてるのは、クリップ名の持ち方がちょっと違うためです。
+    /// </summary>
+    public class FaceMotionBlendShapeNameStore
+    {
+        public FaceMotionBlendShapeNameStore()
+        {
+            BlendShapeNames = new ReadOnlyObservableCollection<string>(_blendShapeNames);
+            var defaultNames = LoadDefaultNames();
+            for (int i = 0; i < defaultNames.Length; i++)
+            {
+                _blendShapeNames.Add(defaultNames[i]);
+            }
+        }
+
+        private readonly ObservableCollection<string> _blendShapeNames = new ObservableCollection<string>();
+        /// <summary> UIに表示するのが妥当と考えられるブレンドシェイプクリップ名の一覧です。 </summary>
+        public ReadOnlyObservableCollection<string> BlendShapeNames { get; }
+
+        //Unityで読み込まれたキャラクターのブレンドシェイプ名の一覧です。
+        //NOTE: この値は標準ブレンドシェイプ名を含んでいてもいなくてもOK。ただし現行動作では標準ブレンドシェイプ名は含まない。
+        private string[] _avatarClipNames = Array.Empty<string>();
+
+        //設定ファイルから読み込んだ設定で使われていたブレンドシェイプ名の一覧。
+        //NOTE: この値に標準ブレンドシェイプ名とそうでないのが混在することがあるが、それはOK
+        private string[] _settingUsedNames = Array.Empty<string>();
+
+        /// <summary>
+        /// ロードされたVRMの標準以外のブレンドシェイプ名を指定して、名前一覧を更新します。
+        /// </summary>
+        /// <param name="avatarBlendShapeNames"></param>
+        public void Refresh(string[] avatarBlendShapeNames)
+        {
+            //なんとなく正格評価しておく(値コピーの方が安心なので…
+            _avatarClipNames = avatarBlendShapeNames.ToArray();
+            RefreshInternal();
+        }
+
+        /// <summary>
+        /// ファイルからロードされたはずの設定を参照し、その中で使われているブレンドシェイプ名を参考にして名前一覧を更新します。
+        /// </summary>
+        /// <param name="neutralClipName"></param>
+        /// <param name="offsetClipName"></param>
+        public void Refresh(string? neutralClipName, string? offsetClipName)
+        {
+            //NOTE: Refreshの挙動上、ここで""だけを2つ入れたりしても大丈夫。
+            _settingUsedNames = new string[] { neutralClipName ?? "", offsetClipName ?? "" };
+            RefreshInternal();
+        }
+
+        private void RefreshInternal()
+        {
+            //理想の並び: デフォルトのやつ一覧、今ロードしたVRMにある名前一覧、(今ロードしたVRMにはないけど)設定で使ってる名前一覧
+            var newNames = LoadDefaultNames().ToList();
+            int defaultSetLength = newNames.Count;
+            foreach (var nameInModel in _avatarClipNames)
+            {
+                if (!newNames.Contains(nameInModel))
+                {
+                    newNames.Add(nameInModel);
+                }
+            }
+
+            foreach (var nameInSetting in _settingUsedNames)
+            {
+                if (!newNames.Contains(nameInSetting))
+                {
+                    newNames.Add(nameInSetting);
+                }
+            }
+
+            var newNameArray = newNames.ToArray();
+
+            //NOTE: ここポイントで、既存要素は消さないよう慎重に並べ替えます(消すとOC<T>の怒りを買ってUI側の要素選択に悪影響が出たりするので…)
+            for (int i = defaultSetLength; i < newNameArray.Length; i++)
+            {
+                if (_blendShapeNames.Contains(newNameArray[i]))
+                {
+                    int currentIndex = _blendShapeNames.IndexOf(newNameArray[i]);
+                    if (currentIndex != i)
+                    {
+                        //もう入ってる値だが、場所を入れ替えたいケース
+                        _blendShapeNames.Move(currentIndex, i);
+                    }
+                }
+                else
+                {
+                    //そもそも入ってないケース
+                    _blendShapeNames.Insert(i, newNameArray[i]);
+                }
+            }
+
+            //OC<T>側のほうが配列が長い場合、ハミ出た分は余計なやつなので消しちゃってOK
+            while (_blendShapeNames.Count > newNameArray.Length)
+            {
+                _blendShapeNames.RemoveAt(newNameArray.Length);
+            }
+        }
+
+        private static string[] LoadDefaultNames()
+        {
+            return new string[]
+            {
+                //「なし」があるのが大事。これによって、条件に合致しても何のブレンドシェイプを起動しない！という事ができる。
+                "",
+                "Joy",
+                "Angry",
+                "Sorrow",
+                "Fun",
+
+                "A",
+                "I",
+                "U",
+                "E",
+                "O",
+
+                "Neutral",
+                "Blink",
+                "Blink_L",
+                "Blink_R",
+
+                "LookUp",
+                "LookDown",
+                "LookLeft",
+                "LookRight",
+            };
+        }
+    }
+
+}
