+++ WPF/VMagicMirrorConfig/Model/Avatar/LoadedAvatarInfo.cs
