+++ WPF/VMagicMirrorConfig/Model/SettingModel/ExternalTrackerSettingModel.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Net;
 using System.Threading.Tasks;
 
 namespace Baku.VMagicMirrorConfig
@@ -135,6 +136,21 @@ namespace Baku.VMagicMirrorConfig
             if (result)
             {
                 EnableExternalTracking.Value = false;
+            }
+        }
+
+        public void RefreshConnectionIfPossible()
+        {
+            //NOTE: 今は選択肢がiFacialMocapのみのため、そこをピンポイントで見に行く
+            if (!EnableExternalTracking.Value || TrackSourceType.Value != ExternalTrackerSetting.TrackSourceIFacialMocap)
+            {
+                return;
+            }
+
+            var ipAddress = IFacialMocapTargetIpAddress.Value;
+            if (IPAddress.TryParse(ipAddress, out _))
+            {
+                NetworkEnvironmentUtils.SendIFacialMocapDataReceiveRequest(ipAddress);
             }
         }
     }
