+++ WPF/VMagicMirrorConfig/View/ControlPanel/AccessoryPanel.xaml
@@ -8,8 +8,10 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              mc:Ignorable="d" 
              d:DesignHeight="850" d:DesignWidth="550"
-             d:DataContext="{d:DesignInstance vm:AccessorySettingViewModel}"
              >
+    <UserControl.DataContext>
+        <vm:AccessorySettingViewModel />
+    </UserControl.DataContext>
     <UserControl.Resources>
         <view:StringNotEmptyToVisibilityConverter x:Key="StringNotEmptyToVisibilityConverter" />
         <view:BooleanToOpacityConverter x:Key="BooleanToOpacityConverter"/>
