+++ WPF/VMagicMirrorConfig/View/ControlPanel/HandTrackingPanel.xaml
@@ -8,8 +8,11 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              mc:Ignorable="d" 
              d:DesignHeight="850" d:DesignWidth="550"
-             d:DataContext="{d:DesignInstance Type={x:Type vm:MotionSettingViewModel}}"
+             d:DataContext="{d:DesignInstance Type={x:Type vm:HandTrackingViewModel}}"
              >
+    <UserControl.DataContext>
+        <vm:HandTrackingViewModel />
+    </UserControl.DataContext>
     <UserControl.Resources>
         <view:StringNotEmptyToVisibilityConverter x:Key="StringNotEmptyToVisibilityConverter" />
         <view:BooleanToOpacityConverter x:Key="BooleanToOpacityConverter"/>
