+++ WPF/VMagicMirrorConfig/View/SettingWindowTabs/EffectSettingPanel.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Baku.VMagicMirrorConfig.View.LightSettingPanel"
+﻿<UserControl x:Class="Baku.VMagicMirrorConfig.View.EffectSettingPanel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -9,8 +9,10 @@
              mc:Ignorable="d"
              d:DesignHeight="1350"
              d:DesignWidth="350"
-             d:DataContext="{d:DesignInstance Type={x:Type vm:LightSettingViewModel}}"
              >
+    <UserControl.DataContext>
+        <vm:EffectSettingViewModel />
+    </UserControl.DataContext>
     <UserControl.Resources>
         <view:WhiteSpaceStringToNullConverter x:Key="WhiteSpaceStringToNullConverter"/>
         <Style TargetType="CheckBox" BasedOn="{StaticResource {x:Type CheckBox}}">
