+++ WPF/VMagicMirrorConfig/View/SettingWindowTabs/EffectSettingPanel.xaml.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows.Controls;
+
+namespace Baku.VMagicMirrorConfig.View
+{
+    public partial class EffectSettingPanel : UserControl
+    {
+        public EffectSettingPanel() => InitializeComponent();
+    }
+}
