--- WPF/VMagicMirrorConfig/View/SettingWindowTabs/FaceMotionSettingPanel.xaml.cs
@@ -1,9 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Baku.VMagicMirrorConfig.View
-{
-    public partial class FaceMotionSettingPanel : UserControl
-    {
-        public FaceMotionSettingPanel() => InitializeComponent();
-    }
-}
