+++ WPF/VMagicMirrorConfig/View/SettingWindowTabs/FaceSettingPanel.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Baku.VMagicMirrorConfig.View.FaceMotionSettingPanel"
+﻿<UserControl x:Class="Baku.VMagicMirrorConfig.View.FaceSettingPanel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -8,10 +8,12 @@
              xmlns:vm="clr-namespace:Baku.VMagicMirrorConfig.ViewModel"
              xmlns:vmm="clr-namespace:Baku.VMagicMirrorConfig"
              mc:Ignorable="d"
-             d:DataContext="{d:DesignInstance Type={x:Type vm:MotionSettingViewModel}}"
              d:DesignWidth="400"
              d:DesignHeight="1200"
              >
+    <UserControl.DataContext>
+        <vm:FaceSettingViewModel />
+    </UserControl.DataContext>
     <UserControl.Resources>
         <view:WhiteSpaceStringToNullConverter x:Key="WhiteSpaceStringToNullConverter"/>
         <view:Lv50Converter x:Key="Lv50Converter"/>
