+++ WPF/VMagicMirrorConfig/View/SettingWindowTabs/FaceSettingPanel.xaml.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows.Controls;
+
+namespace Baku.VMagicMirrorConfig.View
+{
+    public partial class FaceSettingPanel : UserControl
+    {
+        public FaceSettingPanel() => InitializeComponent();
+    }
+}
