--- WPF/VMagicMirrorConfig/View/SettingWindowTabs/LightSettingPanel.xaml.cs
@@ -1,9 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Baku.VMagicMirrorConfig.View
-{
-    public partial class LightSettingPanel : UserControl
-    {
-        public LightSettingPanel() => InitializeComponent();
-    }
-}
