+++ WPF/VMagicMirrorConfig/ViewModel/Infrastructure/ViewModelBase.cs
@@ -1,6 +1,13 @@
-﻿namespace Baku.VMagicMirrorConfig.ViewModel
+﻿using System.ComponentModel;
+using System.Windows;
+
+namespace Baku.VMagicMirrorConfig.ViewModel
 {
     public abstract class ViewModelBase : NotifiableBase
     {
+        private static readonly DependencyObject designerModeCheckObject 
+            = new DependencyObject();
+
+        protected static bool IsInDegignMode => DesignerProperties.GetIsInDesignMode(designerModeCheckObject);
     }
 }
