+++ WPF/VMagicMirrorConfig/ViewModel/Update/UpdateNotificationViewModel.cs
@@ -5,7 +5,6 @@
         public UpdateNotificationViewModel(UpdateCheckResult modelData)
         {
             _modelData = modelData;
-
             CurrentVersion = AppConsts.AppVersion.ToString();
             LatestVersion = modelData.Version.ToString();
             ReleaseDate = string.IsNullOrEmpty(modelData.ReleaseNote.DateString)
